+++ src/bmlib/include/optim/misc/transform_vals.hpp
@@ -1,6 +1,6 @@
 /*################################################################################
   ##
-  ##   Copyright (C) 2011-2018 Keith O'Hara
+  ##   Copyright (C) 2016-2018 Keith O'Hara
   ##
   ##   This file is part of the OptimLib C++ library.
   ##
@@ -18,24 +18,20 @@
  
 /*
  * transform values
- *
- * Keith O'Hara
- * 05/01/2012
- *
- * This version:
- * 08/12/2017
  */
 
 inline
 arma::vec
 transform(const arma::vec& vals_inp, const arma::uvec& bounds_type, const arma::vec& lower_bounds, const arma::vec& upper_bounds)
 {
-    const int n_vals = bounds_type.n_elem;
+    const size_t n_vals = bounds_type.n_elem;
 
     arma::vec vals_trans_out(n_vals);
 
-    for (int i=0; i < n_vals; i++) {
-        switch (bounds_type(i)) {
+    for (size_t i=0; i < n_vals; i++)
+    {
+        switch (bounds_type(i))
+        {
             case 1: // no bounds
                 vals_trans_out(i) = vals_inp(i);
                 break;
@@ -50,7 +46,9 @@ transform(const arma::vec& vals_inp, const arma::uvec& bounds_type, const arma::
                 break;
         }
     }
+
     //
+
     return vals_trans_out;
 }
 
@@ -58,12 +56,14 @@ inline
 arma::vec
 inv_transform(const arma::vec& vals_trans_inp, const arma::uvec& bounds_type, const arma::vec& lower_bounds, const arma::vec& upper_bounds)
 {
-    const int n_vals = bounds_type.n_elem;
+    const size_t n_vals = bounds_type.n_elem;
 
     arma::vec vals_out(n_vals);
 
-    for (int i=0; i < n_vals; i++) {
-        switch (bounds_type(i)) {
+    for (size_t i=0; i < n_vals; i++)
+    {
+        switch (bounds_type(i))
+        {
             case 1: // no bounds
                 vals_out(i) = vals_trans_inp(i);
                 break;
@@ -86,6 +86,8 @@ inv_transform(const arma::vec& vals_trans_inp, const arma::uvec& bounds_type, co
                 break;
         }
     }
+
     //
+    
     return vals_out;
 }
