+++ src/bmlib/include/optim/misc/unit_vec.ipp
@@ -1,6 +1,6 @@
 /*################################################################################
   ##
-  ##   Copyright (C) 2016-2017 Keith O'Hara
+  ##   Copyright (C) 2016-2018 Keith O'Hara
   ##
   ##   This file is part of the OptimLib C++ library.
   ##
@@ -18,14 +18,11 @@
 
 /*
  * unit vector
- *
- * Keith O'Hara
- * 11/28/2014
  */
 
 inline
 arma::vec
-unit_vec(const int j, const int n)
+unit_vec(const size_t j, const size_t n)
 {
     arma::vec ret = arma::zeros(n,1);
     ret(j) = 1;
